requirements.txt
aiogrpc>=1.4
cryptography
flask
googleapis-common-protos
numpy
opencv-python
Pillow
requests